# Improve performance of formatting a project