--- a/src/Juvix/Compiler/Concrete/Translation/FromSource/Data/Context.hs
+++ b/src/Juvix/Compiler/Concrete/Translation/FromSource/Data/Context.hs
@@ -10,3 +10,6 @@ data ParserResult = ParserResult
   }
 
 makeLenses ''ParserResult
+
+getParserResultComments :: ParserResult -> Comments
+getParserResultComments sr = mkComments $ sr ^. resultParserState . parserStateComments