… src/Juvix/Compiler/Concrete/Data/Scope/Base.hs
… src/Juvix/Compiler/Concrete/Language.hs
… src/Juvix/Compiler/Concrete/Translation/FromParsed.hs
… src/Juvix/Compiler/Concrete/Translation/FromParsed/Analysis/Scoping.hs
… src/Juvix/Compiler/Concrete/Translation/FromParsed/Analysis/Scoping/Data/Context.hs
… src/Juvix/Compiler/Concrete/Translation/FromSource.hs
… src/Juvix/Compiler/Internal/Translation/FromConcrete.hs
… src/Juvix/Compiler/Pipeline.hs
… src/Juvix/Compiler/Pipeline/Driver.hs
… src/Juvix/Compiler/Pipeline/DriverParallel.hs
… src/Juvix/Compiler/Pipeline/EntryPoint.hs
… src/Juvix/Compiler/Pipeline/EntryPoint/IO.hs
… src/Juvix/Compiler/Pipeline/JvoCache.hs
… src/Juvix/Compiler/Pipeline/Loader/PathResolver/Base.hs
… src/Juvix/Compiler/Pipeline/Loader/PathResolver/ImportTree/Base.hs
… src/Juvix/Compiler/Pipeline/Loader/PathResolver/Paths.hs
… src/Juvix/Compiler/Pipeline/ModuleInfoCache.hs
… src/Juvix/Compiler/Pipeline/Package/Loader/EvalEff/IO.hs
… src/Juvix/Compiler/Pipeline/Repl.hs
… src/Juvix/Compiler/Pipeline/Run.hs
… src/Juvix/Compiler/Store/Extra.hs
… src/Juvix/Compiler/Store/Language.hs
… src/Juvix/Compiler/Store/Scoped/Language.hs
… src/Juvix/Data.hs
… src/Juvix/Data/Effect.hs
… src/Juvix/Data/Effect/Cache.hs
… src/Juvix/Data/Effect/Forcing.hs
… src/Juvix/Data/ModuleId.hs
… src/Juvix/Data/TopModulePathKey.hs
… src/Juvix/Formatter.hs